refactor(filehandler): Move abstract `FileHandler` into submodule
This makes the `__init__.py` file a lot shorter, and it also allows the
FileHandler implementations to follow the import guidelines to "import
modules, not classes".
